UsersActivationServiceImplTest: adapted tests after changing countByActivationKey() return type.
Correction for 8bf05d9a368b20bbf56922c018450b6023fc08b9 commit.

No functional changes.

+++ src/test/groovy/ru/mystamps/web/service/UsersActivationServiceImplTest.groovy
@@ -167,11 +167,11 @@ class UsersActivationServiceImplTest extends Specification {
 	
 	def "countByActivationKey() should call dao"() {
 		given:
-			usersActivationDao.countByActivationKey(_ as String) >> 2
+			usersActivationDao.countByActivationKey(_ as String) >> 2L
 		when:
-			int result = service.countByActivationKey('0123456789')
+			long result = service.countByActivationKey('0123456789')
 		then:
-			result == 2
+			result == 2L
 	}
 	
 	def "countByActivationKey() should pass activation key to dao"() {
